requirements.txt
anyio==4.7.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.3
cocos2d==0.6.10
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
dash==2.9.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
fonttools==4.42.0
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.2
numpy==1.24.2
packaging==23.1
pandas==1.5.3
Pillow==10.0.0
plotly==5.13.1
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyglet==1.5.27
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.1
python-telegram-bot==21.9
pytz==2022.7.1
requests==2.32.3
rsa==4.9
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
tenacity==8.2.2
tornado==6.4.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.3
